+++ src/commands/add/index.ts
@@ -1,9 +1,9 @@
-import { ArgType } from '../Models/Command.js';
-import { EPermissionLevel } from './../Typings/enums.js';
-import gql, { EmoteSearchFilter, EmoteSet, ListItemAction } from './../SevenTVGQL.js';
-import SevenTVAllowed, { Get7TVUserMod } from './../PreHandlers/7tv.can.modify.js';
-import { extractSeventTVID } from './../tools/regex.js';
-import { registerCommand } from '../controller/Commands/Handler.js';
+import { ArgType } from '../../Models/Command.js';
+import { EPermissionLevel } from './../../Typings/enums.js';
+import gql, { EmoteSearchFilter, EmoteSet, ListItemAction } from './../../SevenTVGQL.js';
+import SevenTVAllowed, { Get7TVUserMod } from './../../PreHandlers/7tv.can.modify.js';
+import { extractSeventTVID } from './../../tools/regex.js';
+import { registerCommand } from '../../controller/Commands/Handler.js';
 
 type PreHandlers = {
 	SevenTV: Get7TVUserMod;
