+++ src/commands/editor/index.ts
@@ -1,8 +1,8 @@
-import { EPermissionLevel } from './../Typings/enums.js';
-import gql, { UserEditorPermissions } from '../SevenTVGQL.js';
-import SevenTVAllowed, { Get7TVUserMod } from './../PreHandlers/7tv.can.modify.js';
-import PreHandlers from './../PreHandlers/index.js';
-import { registerCommand } from '../controller/Commands/Handler.js';
+import { EPermissionLevel } from './../../Typings/enums.js';
+import gql, { UserEditorPermissions } from '../../SevenTVGQL.js';
+import SevenTVAllowed, { Get7TVUserMod } from './../../PreHandlers/7tv.can.modify.js';
+import PreHandlers from './../../PreHandlers/index.js';
+import { registerCommand } from '../../controller/Commands/Handler.js';
 
 type PreHandlers = {
 	SevenTV: Get7TVUserMod;
