+++ src/commands/hint/index.ts
@@ -1,5 +1,6 @@
-import { registerCommand } from '../controller/Commands/Handler.js';
-import { EPermissionLevel } from './../Typings/enums.js';
+import { registerCommand } from '../../controller/Commands/Handler.js';
+import { EPermissionLevel } from './../../Typings/enums.js';
+
 registerCommand({
 	Name: 'hint',
 	Ping: false,
